Move subfolders containing included files to src subfolder
Recent versions of the Arduino IDE only support recursive compilation within the src subfolder of the sketch folder (https://github.com/arduino/arduino-builder/pull/148). Because of this, compilation of BMSino would fail with a lot of "undefined reference" errors, which is solved by moving the files to the src subfolder and updating the includes.